separation of main functionality
