+++ platformio.ini
@@ -47,8 +47,9 @@ lib_deps = throwtheswitch/Unity@^2.5.2
     https://github.com/meekrosoft/fff.git
 build_flags = -D WINDOWS_TEST=1 -D TEMP_HUMID
 
-[env:win_test_light]
+[env:win_test_main]
 platform = native
 lib_deps = throwtheswitch/Unity@^2.5.2
     https://github.com/DaveGamble/cJSON.git
-build_flags = -DWINDOWS_TEST -DEXCLUDE_UART
+    https://github.com/meekrosoft/fff.git
+build_flags = -D WINDOWS_TEST=1 -D MAIN
