+++ test/mocks.h
@@ -15,3 +15,7 @@ DECLARE_FAKE_VOID_FUNC(display_int, int16_t);
 #ifdef WEATHER
 DECLARE_FAKE_VALUE_FUNC(TempHumid, temperature_humidity_get_combined_values);
 #endif
+#ifdef MAIN
+DECLARE_FAKE_VOID_FUNC(wifi_init);
+DECLARE_FAKE_VOID_FUNC(weather_init);
+#endif
